Move DirectedSphereExclusion
The DirectedSphereExclusion is a distance-based method (similar to OptiSim).
So, it should be moved to methods.dissimilarity module. See #133
